+++ R/Future-class.R
@@ -68,7 +68,7 @@
 #' @param lazy If FALSE (default), the future is resolved
 #' eagerly (starting immediately), otherwise not.
 #'
-#' @param local If TRUE, the expression is evaluated such that
+#' @param local (deprecated) If TRUE, the expression is evaluated such that
 #' all assignments are done to local temporary environment, otherwise
 #' the assignments are done to the global environment of the \R process
 #' evaluating the future.
@@ -142,11 +142,7 @@ Future <- function(expr = NULL, envir = parent.frame(), substitute = TRUE, stdou
   args <- list(...)
 
   if (!local && !isTRUE(args[["persistent"]])) {
-    ## Note, this warning will be muffled for UniprocessFuture:s when R option
-    ## 'nwarnings' is reset at the end of evaluating the future expression,
-    ## e.g. local({ warning("boom"); options(nwarnings = 50L) }).
-    ## /HB 2022-04-27
-    .Deprecated(msg = "Using 'local = FALSE' for a future is deprecated in future (>= 1.20.0) and will soon be defunct and produce an error.", package = .packageName)
+    .Defunct(msg = "Using 'local = FALSE' for a future is defunct in future (>= 1.28.0)", package = .packageName)
   }
 
   core <- new.env(parent = emptyenv())